requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
greenlet==3.0.3
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF Generation
weasyprint==60.2

# AI Integration
openai==1.12.0

# Cloud Storage
boto3==1.34.34

# Utilities
python-dotenv==1.0.0
